fix: Add sidebar switch icon and adjust UI elements
- Added new sidebar-switch-symbolic icon to resources
- Reduced expand button icon size in FileManagerWindow
- Updated sidebar switch icon to use the new symbolic icon

This commit improves the visual consistency and adds a new symbolic icon for sidebar interactions.

Log: replace sidebar switch icon
Bug: https://pms.uniontech.com/bug-view-304309.html
