+++ src/dfm-base/widgets/dfmcustombuttons/customiconbutton.cpp
@@ -1,8 +1,9 @@
-// SPDX-FileCopyrightText: 2023 UnionTech Software Technology Co., Ltd.
+// SPDX-FileCopyrightText: 2024 UnionTech Software Technology Co., Ltd.
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-#include "custombutton.h"
+#include "customiconbutton.h"
+#include <dfm-base/dfm_base_global.h>
 
 #include <DGuiApplicationHelper>
 #include <DStylePainter>
@@ -16,6 +17,7 @@
 #include <QPainterPath>
 
 DWIDGET_USE_NAMESPACE
+DFMBASE_USE_NAMESPACE
 
 CustomDIconButton::CustomDIconButton(QWidget *parent)
     : DIconButton(parent)
@@ -53,29 +55,4 @@ void CustomDIconButton::paintEvent(QPaintEvent *event)
 
     DStyleHelper dstyle(style());
     dstyle.drawControl(DStyle::CE_IconButton, &opt, &painter, this);
-}
-
-CustomDToolButton::CustomDToolButton(QWidget *parent)
-    : DToolButton(parent)
-{
-}
-
-void CustomDToolButton::initStyleOption(QStyleOptionToolButton *option) const
-{
-    DToolButton::initStyleOption(option);
-    if (underMouse()) {
-        bool isDarkTheme = DGuiApplicationHelper::instance()->themeType() == DGuiApplicationHelper::DarkType;
-        QColor hoverColor = isDarkTheme ? QColor(255, 255, 255, 15)
-                                        : QColor(0, 0, 0, 26);
-        option->palette.setColor(QPalette::Button, hoverColor);
-    }
-}
-
-void CustomDToolButton::paintEvent(QPaintEvent *event)
-{
-    Q_UNUSED(event)
-    QStylePainter p(this);
-    QStyleOptionToolButton opt;
-    initStyleOption(&opt);
-    p.drawComplexControl(QStyle::CC_ToolButton, opt);
 }
