+++ src/dfm-base/widgets/dfmcustombuttons/customiconbutton.h
@@ -1,13 +1,15 @@
-// SPDX-FileCopyrightText: 2023 UnionTech Software Technology Co., Ltd.
+// SPDX-FileCopyrightText: 2024 UnionTech Software Technology Co., Ltd.
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-#ifndef CUSTOMBUTTON_H
-#define CUSTOMBUTTON_H
+#ifndef CUSTOMICONBUTTON_H
+#define CUSTOMICONBUTTON_H
 
 #include <DIconButton>
 #include <DToolButton>
 #include <DStyle>
+
+namespace dfmbase {
 
 class CustomDIconButton : public DTK_NAMESPACE::Widget::DIconButton
 {
@@ -20,15 +22,6 @@ protected:
     void paintEvent(QPaintEvent *event) override;
 };
 
-class CustomDToolButton : public DTK_NAMESPACE::Widget::DToolButton
-{
-    Q_OBJECT
-public:
-    explicit CustomDToolButton(QWidget *parent = nullptr);
+}
 
-protected:
-    void paintEvent(QPaintEvent *event) override;
-    void initStyleOption(QStyleOptionToolButton *option) const override;
-};
-
-#endif   // CUSTOMBUTTON_H
+#endif   // CUSTOMICONBUTTON_H
